feat(magiclink-auth-api): add docker file. Update docker-compose. Small code refactoring
