+++ packages/magiclink-auth-service-api/.eslintignore
@@ -1,0 +1,2 @@
+node_modules
+src/public/js
