+++ packages/magiclink-auth-service-api/src/app.js
@@ -5,11 +5,8 @@ const uuid = require('uuid');
 const logger = require('./util/logger');
 const routes = require('./routes');
 const apiErrorHandler = require('./error/apiErrorHandler');
-const passport = require('passport');
-const jwtAuthStrategy = require('./config/jwtAuthStrategy.js');
 
 const app = express();
-jwtAuthStrategy.use(passport);
 
 app
   .use(
@@ -20,7 +17,6 @@ app
   )
   .use(bodyParser.json())
   .use('/', routes)
-  .use(passport.initialize())
   .use(apiErrorHandler);
 
 module.exports = app;
