+++ packages/magiclink-auth-service-api/src/middleware/validate-magiclink-payload.js
@@ -5,7 +5,6 @@ const logger = require('../util/logger');
 module.exports = async (req, res, next) => {
   try {
     req.body = await magicLinkDataValidator(req.body);
-
     logger.debug('Magic link request body is valid.');
     return next();
   } catch (err) {
