+++ packages/magiclink-auth-service-api/src/routes/magiclink.js
@@ -1,11 +1,11 @@
 const express = require('express');
-const validateMagicLink = require('../middleware/authenticate');
+const authenticate = require('../middleware/authenticate');
 const validateMagicLinkPayload = require('../middleware/validate-magiclink-payload');
 const magicLinkController = require('../controllers/magiclink');
 const router = express.Router();
 
 router.post('/magiclink', validateMagicLinkPayload, magicLinkController.create);
-router.get('/magiclink/:magiclink', validateMagicLink, magicLinkController.login);
+router.get('/magiclink/:magiclink', authenticate, magicLinkController.login);
 
 router.get('/hello', (req, res) => {
   return res.status(200).send('Hello world');
