+++ packages/magiclink-auth-service-api/src/service/error/InvalidTokenError.js
