+++ packages/magiclink-auth-service-api/src/util/crypto.js
@@ -2,7 +2,7 @@ const crypto = require('crypto');
 const config = require('../config');
 const algorithm = 'aes-256-ctr';
 const iv = crypto.randomBytes(16);
-const secretKey = config.cryptoKey;
+const secretKey = config.dataEncryptionKey;
 
 const encrypt = (text) => {
   const cipher = crypto.createCipheriv(algorithm, secretKey, iv);
