+++ packages/magiclink-auth-service-api/src/util/jwt.js
@@ -2,6 +2,5 @@ const jwt = require('jsonwebtoken');
 const config = require('../config');
 
 module.exports.sign = (payload) => {
-  // TODO sign with RSA SHA256
-  return jwt.sign(payload, config.jwtKey);
+  return jwt.sign(payload, config.jwtSigningKey);
 };
